requirements.txt
django-cors-headers==2.4.0
django-eav2==0.11.1
django==2.1
djangorestframework==3.8.2
psycopg2==2.7.5
pytz==2018.5
